requirements.txt
pyyaml==5.4.1
rich==9.11.1